+++ packages/platform/web/apps/complete-application/src/app/projects/[project_id]/analytics/page.tsx
@@ -1,6 +1,10 @@
+import ComingSoon from '@/components/coming-soon'
 import Analytics from '@/features/projects/analytics'
 export default function page() {
   return (
-    <div><Analytics /></div>
+    <div>
+      <ComingSoon/>
+      {/* <Analytics /> */}
+      </div>
   )
 }
