+++ packages/platform/web/apps/complete-application/src/app/projects/[project_id]/hub-listing/layout.tsx
@@ -1,0 +1,11 @@
+export default function HubListingLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <div>{children}</div>
+    </>
+  );
+}
